+++ examples/taskqueue/sampletasks/sampletasks.py
@@ -1,7 +1,9 @@
-from pulsar.apps.tasks import Job
+import time
+
+from pulsar.apps import tasks
 
 
-class RunPyCode(Job):
+class RunPyCode(tasks.Job):
     '''execute python code in *code*. There must be a *task_function*
 function defined.'''
     def __call__(self, consumer, code, *args, **kwargs):
@@ -12,7 +14,15 @@ function defined.'''
         return func(*args,**kwargs)
         
 
-class Addition(Job):
+class Addition(tasks.Job):
     
     def __call__(self, consumer, a, b):
         return a+b
+    
+    
+class NotOverLap(tasks.Job):
+    can_overlap = False
+    
+    def __call__(self, consumer, lag, *args, **kwargs):
+        sleep(lag)
+        return 'OK'