+++ pulsar/apps/tasks/scheduler.py
@@ -163,8 +163,10 @@ and task scheduling."""
 :meth:`make_request` and check that the task can be inserted into the queue.'''
         task = self.make_request(job_name, targs, tkwargs, **kwargs)
         if task.needs_queuing():
+            task._queued = True
             monitor.put(task)
         else:
+            task._queued = False
             self.log.info('task {0} already requested. Abort request.'\
                                   .format(task))
         return task