+++ CHANGELOG.rst
@@ -5,12 +5,12 @@ Version Master
 * Added :class:`pulsar.Mailbox` classes for handling message passing between actors.
 * Added :mod:`pulsar.apps.ws`, an asynchronous websocket application for pulsar.
 * Asynchronous applications.
-* Create the :mod:`pulsar.net` module for stream primitive.
+* Created the :mod:`pulsar.net` module for internet primitive.
 * Added a windows service wrapper class for using pulsar with windows services.
 * Removed the `worker` module.
 * Moved `http.rpc` module to `apps`.
-* Introduced context manager for `apps.tasks` to handle logs and exceptions.
-* **37 regression tests**
+* Introduced context manager for `pulsar.apps.tasks` to handle logs and exceptions.
+* **43 regression tests**
 
 Version 0.1.0 - 2001-Aug-24
 =======================================